+++ set-env-local.sh
@@ -1,3 +1,6 @@
 export CELERY_BROKER_URL='amqp://roettger:secert7$@localhost:5672/topobank'
 export FIREFOX_BINARY_PATH=`which firefox`
 export GECKODRIVER_PATH=`which geckodriver`
+export EMAIL_HOST=localhost
+export EMAIL_PORT=8025
+