+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog for *TopoBank*
 
+## 0.9.4 (2020-11-23)
+
+- added new analysis function "RMS Values" (#445)
+- autoscroll in select tab which moves selected 
+  row into view (#544)
+- improved bandwidth plot, removed labels
+  and added info box on hover with name and 
+  thumbnail of the corresponding topography (#535)
+- enabled custom sorting of sharing info table (#300)
+- usage statistics can be sent by email (#550)  
+- fixed wrong select tab state in session (#532)
+- renamed button "Properties" to "View" (#537)
+- on select tab, only showing first line of 
+  descriptions on load, rest expandable by 
+  button press (#540)
+- fixed ordering by surface name on publication tab
+
 ## 0.9.3 (2020-11-12)
 
 - fixed XSS issues